create print_p

+++ main.h
@@ -18,6 +18,6 @@ int print_char(int c);
 int print_hex(unsigned int num, char c);
 int print_octal(unsigned int num);
 int print_binary(unsigned int num);
-int print_p(unsigned int num);
+int print_p(unsigned long long num);
 
 #endif
